nros_core::error

Trait TakeFailedAsNone

pub trait TakeFailedAsNone {
    type T;

    // Required method
    fn take_failed_as_none(self) -> Result<Option<Self::T>, NanoRosError>;
}

A helper trait to convert take failures to None

This is useful when you want to distinguish between “no data available” (returns None) and actual errors (returns Err).

Required Associated Types

type T

The value type

Required Methods

fn take_failed_as_none(self) -> Result<Option<Self::T>, NanoRosError>

If the take failed, return Ok(None). Otherwise return Ok(Some(value)).